JNM Refactor: Moved converters into own package
