+++ JavaIntegration/src/main/java/org/openzen/zencode/java/module/converters/JavaNativeMemberConverter.java
@@ -1,6 +1,7 @@
-package org.openzen.zencode.java.module;
+package org.openzen.zencode.java.module.converters;
 
 import org.openzen.zencode.java.ZenCodeType;
+import org.openzen.zencode.java.module.TypeVariableContext;
 import org.openzen.zencode.shared.CodePosition;
 import org.openzen.zencode.shared.CompileException;
 import org.openzen.zencode.shared.LiteralSourceFile;
